backend_python/requirements.txt
openai
python-dotenv
fastapi[standard]